extract useReducer to custom hook
