+++ app/javascript/components/tooltips/Tooltip.tsx
@@ -1,89 +1,7 @@
-import React, { useEffect, useReducer, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { usePopper } from 'react-popper'
 import { useQuery } from 'react-query'
-
-type TooltipState = {
-  requestHover: boolean
-  requestFocus: boolean
-  hover: boolean
-  focus: boolean
-  showState: ShowState
-}
-
-type ShowState = 'hidden' | 'loading' | 'invisible' | 'visible' | 'error'
-
-type TooltipAction =
-  | ErrorAction
-  | RequestShowFromRefHoverAction
-  | RequestHideFromRefHoverAction
-  | RequestShowFromRefFocusAction
-  | RequestHideFromRefFocusAction
-  | RequestShowFromHoverAction
-  | RequestHideFromHoverAction
-  | RequestShowFromFocusAction
-  | RequestHideFromFocusAction
-  | LoadedAction
-  | ShowAction
-  | HideAction
-
-type ErrorAction = {
-  type: 'error'
-}
-
-type RequestShowFromRefHoverAction = {
-  type: 'request-show-from-ref-hover'
-}
-
-type RequestHideFromRefHoverAction = {
-  type: 'request-hide-from-ref-hover'
-}
-
-type RequestShowFromRefFocusAction = {
-  type: 'request-show-from-ref-focus'
-}
-
-type RequestHideFromRefFocusAction = {
-  type: 'request-hide-from-ref-focus'
-}
-
-type RequestShowFromHoverAction = {
-  type: 'request-show-hover'
-}
-
-type RequestHideFromHoverAction = {
-  type: 'request-hide-hover'
-}
-
-type RequestShowFromFocusAction = {
-  type: 'request-show-focus'
-}
-
-type RequestHideFromFocusAction = {
-  type: 'request-hide-focus'
-}
-
-type LoadedAction = {
-  type: 'loaded'
-}
-
-type ShowAction = {
-  type: 'show'
-}
-
-type HideAction = {
-  type: 'hide'
-}
-
-type DispatchAction = {
-  id: string
-  action: TooltipAction
-}
-
-type OpenTooltip = [string, React.Dispatch<unknown>]
-
-interface ShowReducer {
-  (state: TooltipState, payload: DispatchAction): TooltipState
-}
+import { useStatefulTooltip } from '../../hooks/use-stateful-tooltip'
 
 interface TooltipProps {
   id: string
@@ -92,137 +10,6 @@ interface TooltipProps {
   contentEndpoint: string
   hoverRequestToShow: boolean
   focusRequestToShow: boolean
-}
-
-function initialTooltipState(): TooltipState {
-  return {
-    requestHover: false,
-    requestFocus: false,
-    hover: false,
-    focus: false,
-    showState: 'hidden',
-  }
-}
-
-/**
- * OpenTooltips maintains an array of open tooltips
- * Only 1 should be active at once, but given the chance of race conditions
- * A array is used to track potentially multiple open
- */
-
-const openTooltips: OpenTooltip[] = []
-
-const indexOfOpenTooltip = (id: string) =>
-  openTooltips.findIndex(([openId]) => openId === id)
-
-const hasOpenTooltip = (id: string) => {
-  return indexOfOpenTooltip(id) === -1
-    ? openTooltips.length > 0
-    : openTooltips.length > 1
-}
-
-const addOpenTooltip = (id: string, dispatcher: React.Dispatch<unknown>) => {
-  if (openTooltips.find(([openId]) => openId === id)) {
-    return
-  }
-
-  openTooltips.push([id, dispatcher])
-}
-
-const removeOpenTooltip = (id: string) => {
-  const index = indexOfOpenTooltip(id)
-  if (index !== -1) {
-    openTooltips.splice(index, 1)
-  }
-}
-
-const closeOtherOpenTooltips = (id: string) => {
-  openTooltips.forEach(([openId, openDispatch]) => {
-    if (openId !== id) {
-      openDispatch({
-        id: openId,
-        action: {
-          type: 'hide',
-        },
-      })
-    }
-  })
-}
-
-const shouldShow = (state: TooltipState): boolean => {
-  return state.focus || state.hover || state.requestFocus || state.requestHover
-}
-
-const handleShowRequest = (state: TooltipState): ShowState => {
-  return shouldShow(state) && state.showState === 'hidden'
-    ? 'loading'
-    : state.showState
-}
-
-const handleHideRequest = (state: TooltipState): ShowState => {
-  return !shouldShow(state) ? 'hidden' : state.showState
-}
-
-/**
- * tooltipReducer
- * This serves as a state reducer for the <Tooltip /> Component for use with
- * the useReducer hook.
- */
-const tooltipReducer: ShowReducer = function (state, body) {
-  const { id, action } = body
-  const nextState: TooltipState = { ...state }
-
-  // Use action to update the next state
-  switch (action.type) {
-    case 'request-show-from-ref-hover':
-      nextState.requestHover = true && !hasOpenTooltip(id)
-      nextState.showState = handleShowRequest(nextState)
-      break
-    case 'request-hide-from-ref-hover':
-      nextState.requestHover = false
-      nextState.showState = handleHideRequest(nextState)
-      break
-    case 'request-show-from-ref-focus':
-      nextState.requestFocus = true
-      nextState.showState = handleShowRequest(nextState)
-      break
-    case 'request-hide-from-ref-focus':
-      nextState.requestFocus = false
-      nextState.showState = handleHideRequest(nextState)
-      break
-    case 'request-show-hover':
-      nextState.hover = true && !hasOpenTooltip(id)
-      nextState.showState = handleShowRequest(nextState)
-      break
-    case 'request-hide-hover':
-      nextState.hover = false
-      nextState.showState = handleHideRequest(nextState)
-      break
-    case 'request-show-focus':
-      nextState.focus = true
-      nextState.showState = handleShowRequest(nextState)
-      break
-    case 'request-hide-focus':
-      nextState.focus = false
-      nextState.showState = handleHideRequest(nextState)
-      break
-    case 'loaded':
-      nextState.showState = 'invisible'
-      break
-    case 'show':
-      nextState.showState = 'visible'
-      break
-    case 'hide':
-      nextState.showState = 'hidden'
-      break
-    case 'error':
-      nextState.showState = 'error'
-      break
-    default:
-      throw new Error('unhandled tooltip reducer action')
-  }
-
-  return nextState
 }
 
 export const Tooltip = ({
@@ -235,11 +22,13 @@ export const Tooltip = ({
 }: TooltipProps): JSX.Element | null => {
   const [tooltipElement, setTooltipElement] = useState<HTMLElement | null>(null)
 
-  const [{ showState }, dispatch] = useReducer(
-    tooltipReducer,
-    null,
-    initialTooltipState
-  )
+  const {
+    state: { showState },
+    dispatch,
+    addOpenTooltip,
+    removeOpenTooltip,
+    closeOtherOpenTooltips,
+  } = useStatefulTooltip()
 
   const popper = usePopper(referenceElement, tooltipElement, {
     placement: 'bottom',
@@ -295,7 +84,7 @@ export const Tooltip = ({
     return () => {
       clearTimeout(timeoutRef)
     }
-  }, [hoverRequestToShow, id])
+  }, [dispatch, hoverRequestToShow, id])
 
   // useEffect for responding to the reference element's request to show when focused
   useEffect(() => {
@@ -323,7 +112,7 @@ export const Tooltip = ({
     return () => {
       clearTimeout(timeoutRef)
     }
-  }, [focusRequestToShow, id])
+  }, [dispatch, focusRequestToShow, id])
 
   // useEffect for responding to the tooltips internal state
   useEffect(() => {
@@ -376,7 +165,16 @@ export const Tooltip = ({
     //     },
     //   })
     // }
-  }, [showState, id, isLoading, isError])
+  }, [
+    showState,
+    id,
+    isLoading,
+    isError,
+    dispatch,
+    removeOpenTooltip,
+    addOpenTooltip,
+    closeOtherOpenTooltips,
+  ])
 
   // short-circuit return null if not ready to display the tooltip
   if ((showState !== 'visible' && showState !== 'invisible') || !htmlContent) {
